+++ pycricinfo/source_models/pages/series.py
@@ -1,8 +1,15 @@
-from pydantic import BaseModel
+from typing import Optional
+
+from pydantic import BaseModel, Field
 
 
-class Series(BaseModel):
+class MatchSeries(BaseModel):
     title: str
     id: str
     link: str
     summary_url: str
+
+
+class MatchType(BaseModel):
+    name: str
+    series: Optional[list[MatchSeries]] = Field(default_factory=list)